@@ -3,6 +3,36 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
+    <parent>
+        <groupId>de.awtools</groupId>
+        <artifactId>master-pom</artifactId>
+        <version>14-java21</version>
+    </parent>
+
+    <artifactId>registrationservice</artifactId>
+    <version>0.4.0-SNAPSHOT</version>
+    <name>Registration REST API project</name>
+    <packaging>war</packaging>
+
+    <properties>
+        <slf4j.version>2.0.9</slf4j.version>
+        <org.springframework.version>6.0.12</org.springframework.version>
+        <org.springframework.spring-data-jpa.version>3.2.1</org.springframework.spring-data-jpa.version>
+        <org.hibernate.version>6.3.1.Final</org.hibernate.version>
+        <com.fasterxml.jackson.core.jackson-annotations.version>2.15.2</com.fasterxml.jackson.core.jackson-annotations.version>
+        <com.fasterxml.jackson.core.jackson-core.version>2.15.2</com.fasterxml.jackson.core.jackson-core.version>
+        <com.fasterxml.jackson.core.jackson-databind.version>2.15.2</com.fasterxml.jackson.core.jackson-databind.version>
+        <!--
+        <betoffice.storage.version>2.8.1</betoffice.storage.version>
+        <betoffice.openligadb.version>1.6.4</betoffice.openligadb.version>
+        <betoffice.jweb-misc.version>1.3.1</betoffice.jweb-misc.version>
+        <betoffice.jweb.version>1.3.1</betoffice.jweb.version>
+        -->
+        <springfox.version>3.0.0</springfox.version>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+    </properties>
+
     <build>
         <pluginManagement>
             <plugins>
