@@ -2,10 +2,10 @@ package de.awtools.registration.cookie;
 
 import java.time.LocalDateTime;
 
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.validation.Valid;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -18,7 +18,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import de.awtools.registration.HttpConst;
 import de.awtools.registration.time.DateTimeJson;
-import io.swagger.annotations.ApiOperation;
 
 /**
  * The controller to store cookie confirmation or cancellation.
@@ -41,7 +40,7 @@ public class CookieController {
      *
      * @return Web-Service reachable?
      */
-    @ApiOperation(value = "ping", nickname = "ping", response = DateTimeJson.class, notes = "Ping this service. Is it reachable?")
+    // @ApiOperation(value = "ping", nickname = "ping", response = DateTimeJson.class, notes = "Ping this service. Is it reachable?")
     @CrossOrigin
     @GetMapping(path = "/ping", produces = HttpConst.JSON_UTF_8)
     public DateTimeJson ping() {
@@ -49,7 +48,7 @@ public class CookieController {
         return dateTimeJson.setDateTime(LocalDateTime.now());
     }
 
-    @ApiOperation(value = "confirmCookie", nickname = "Confirm Cookie", response = DateTimeJson.class, notes = "The user confirmed a cookie.")
+    // @ApiOperation(value = "confirmCookie", nickname = "Confirm Cookie", response = DateTimeJson.class, notes = "The user confirmed a cookie.")
     @CrossOrigin
     @PostMapping(path = "confirmCookie", headers = {
             HttpConst.CONTENT_TYPE }, produces = HttpConst.JSON_UTF_8)
