@@ -4,9 +4,10 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -36,11 +37,16 @@ public class CookieControllerTest {
 
     @BeforeEach
     public void setup() {
-        MockitoAnnotations.initMocks(this);
+        MockitoAnnotations.openMocks(this);
         mockMvc = MockMvcBuilders
                 .standaloneSetup(new CookieController(cookieService))
                 .setHandlerExceptionResolvers(new MyExceptionResolver())
                 .build();
+    }
+
+    @AfterEach
+    public void teardown() throws Exception {
+        MockitoAnnotations.openMocks(this).close();
     }
 
     @Test
@@ -51,9 +57,8 @@ public class CookieControllerTest {
     private static class MyExceptionResolver implements HandlerExceptionResolver {
 
         @Override
-        public ModelAndView resolveException(HttpServletRequest request, HttpServletResponse response, Object handler,
-                Exception ex) {
-
+        public ModelAndView resolveException(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) {
+            // TODO Auto-generated method stub
             return null;
         }
 
