@@ -8,11 +8,13 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import javax.servlet.ServletContext;
+import jakarta.servlet.ServletContext;
 
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.mock.web.MockServletContext;
@@ -52,12 +54,16 @@ public class ValidationTest {
                 .build();
     }
 
+    @AfterEach
+    public void teardown() throws Exception {
+        MockitoAnnotations.openMocks(this).close();
+    }
+
     @Test
     public void ping() throws Exception {
         ServletContext context = webAppContext.getServletContext();
 
-        assertThat(context).isNotNull();
-        assertThat(context).isInstanceOf(MockServletContext.class);
+        assertThat(context).isNotNull().isInstanceOf(MockServletContext.class);
         assertThat(context.getServletContextName())
                 .isEqualTo("MockServletContext");
         assertThat(webAppContext.getBean(RegistrationController.class))
